# PERF: Decision trees: improve prefs by ~20% with very simple changes