I was doing some profiling with py-spy (amazing tool btw!), and I noticed something weird. When I dived in, I quickly found a **very-low hanging fruit for optimization**.

#### Reference Issues/PRs

None

#### Motivation (what is the weird thing I noticed)

In the flame graph produced by py-spy on the script below.

I noticed something weird: `partitioner.next_p` was taking ~8% of the runtime out of 80% taken by the tree building. It was very unexpected because `partitioner.next_p` does almost nothing and does sequential memory access:

<img width="1826" height="146" alt="vlcsnap-2025-09-14-16h09m02s847" src="https://github.com/user-attachments/assets/222acf5c-9175-4184-8178-b77a3e020bc7" />

After investigation, I determined this comes from this line:
`float32_t[::1] feature_values = self.feature_values`

According to ChatGPT (I'm by no mean a Cython expert so :sweat_smile: ), it's because float32_t[::1] is a not so light C-struct under-the-hood, so making a copy of it in a function that's otherwise super light can make the said function almost 10x slower...

#### Changes:

So I removed all patterns `type[::1] x = self.x` for all light methods I'm aware of in the tree module (and replaced `x` by `self.x`). This leads to this new flame graph:

<img width="1826" height="146" alt="after_changes" src="https://github.com/user-attachments/assets/7e52d737-13cc-44d0-ac82-736e7885f9b0" />

Where indeed `next_p` is much faster! And `criterion.update` to!

#### Benchmarks:

Using the script below (with d=20), on my machine:

Prior to changes: regression: 1.23s | classif: 1.37s
After the changes: regression: 1.00s | classif: 1.16s

It's a ~20% speed-up

The speed increase might be less in different configurations (smaller d for instance) and sometimes more (with nans for instance). 

#### The script:

```python
import argparse
from time import perf_counter
import numpy
from sklearn.tree import DecisionTreeRegressor, DecisionTreeClassifier

if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("--clf", action="store_true", help="Use classifier instead of regressor")
    parser.add_argument("--d", type=int, default=20, help="Number of features")
    args = parser.parse_args()
    clf = args.clf
    d = args.d
    n = 3_000_000 // d
    n_fit = 10
    dt = 0
    for _ in range(n_fit):
        X = numpy.random.rand(n, d)
        y = numpy.random.rand(n) + X.sum(axis=1)
        if clf:
            # 3 classes:
            y = (y < numpy.quantile(y, 0.3)) + (y < numpy.quantile(y, 0.6))
        Tree = DecisionTreeClassifier if clf else DecisionTreeRegressor
        t = perf_counter()
        tree = Tree(max_depth=4, max_features=d).fit(X, y)
        dt += perf_counter() - t
    print(dt / n_fit)
```

You can run py-spy with `py-spy record -o profile.svg -n -- python script.py`. Note the `-n` for profiling "native code" (i.e. Cython code in this case).

#### Additional note

It might be a pattern present in other parts of sklearn. So it's worth communicating to other Cython mainteners.

Note that in "big" functions (i.e. not O(1)), this pattern is likely worth-it to avoid the indirection cost of doing `self.x`.