@adam2392 2025-09-14
Oh that's interesting. I always assumed `cdef float32[:] blah = self.blah` just makes a view of the existing memoryview, and/or Cython would be smart to just in-line replace any `blah` references with `self.blah`...

Pinging @scikit-learn/core-devs for more visibility.