@GaelVaroquaux 2025-09-14
That's very cool!

@cakedev0 , can you run also the benchmarking script for random-forests? It might have sped them up (and they are a very very useful learner).