@cakedev0 2025-09-14
@adam2392 

>  just makes a view of the existing memoryview

It's kinda what happens I think. But actually, because it's not just plain C pointers, it seems it takes a few dozens/hundreds of CPU cycles to complete which in comparison to a fast O(1) function might be a very signigicant overhead.

Or maybe the underlying behavior of `self.blah[i]` and `blah[i]` is different, it's what ChatGPT thinks...