@cakedev0 2025-09-14
Benchmark with random forest using the same script, but with random forests and n_jobs=4, n_estimators=16.

Prior: regression: 6.93s | classif: 7.48s
After: regression: 5.16s | classif: 5.96s

It's... even more significant! :rocket: 25% faster.

But I only have my laptop, it gets hot real quick and the first few fits are much faster. I took the average of the last 5 fits, because it looked much more stable from one run to another.