@adam2392 2025-09-14
I suggest we merge https://github.com/scikit-learn/scikit-learn/pull/32100, and then merge in main here, and then re-check CIs here, and then merge (pending everything looks good still)