@GaelVaroquaux 2025-09-14
> I'm surprised the code compiled with that comma...

In normal Python it would create a tuple. Hence the above change would be changing the type of the rhs of the assignment. I'd say that this change should be checked.