@adam2392 2025-09-14
> I'm surprised the code compiled with that comma...
> In normal Python it would create a tuple. Hence the above change would be changing the type of the rhs of the assignment. I'd say that this change should be checked.

Turns out Cython is a pretty smart compiler… Adding these lines to see what is there:

```
        with gil:
            print('inside sparse partitioner')
            print(indptr_start)
            print(type(indptr_start))

# outputs
inside sparse partitioner
0
<class 'int'>
```

Though it does make me wonder if there was some code that was getting generated differently. Turns out that Cython treats the "comma" as declaring multiple variables at a time, but since it only defines one variable, it gets ignored. See the  `_partitioner.c` file that gets generated by Cython

```C++
  /* "sklearn/tree/_partitioner.pyx":492
 *         cdef intp_t[::1] samples = self.samples
 *         cdef float32_t[::1] feature_values = self.feature_values
 *         cdef intp_t indptr_start = self.X_indptr[feature],             # <<<<<<<<<<<<<<
 * 
 *         with gil:
*/
  __pyx_t_3 = __pyx_v_feature;
  __pyx_v_indptr_start = (*((__pyx_t_7sklearn_5utils_9_typedefs_int32_t const  *) ( /* dim=0 */ ((char *) (((__pyx_t_7sklearn_5utils_9_typedefs_int32_t const  *) __pyx_v_self->X_indptr.data) + __pyx_t_3)) )));
```