@da-woods 2025-09-15
Just to point out: this only an improvement because you're set `initializedcheck=False`.  Without that directive memoryview attributes on `cdef classes` are usually assumed to be "maybe None" while local memoryview variables are checked on assignment and then assumed not to be None.

Beyond that, the change is dropping two bits of reference counting and a copy of a moderately sized struct per assignment, so I'm surprised it makes much difference. But it is definitely removing a bit of work.